requirements.txt
# 핵심 라이브러리
duckdb==0.10.0
pandas==2.2.0
numpy==1.26.0
pyarrow==15.0.0
scipy==1.12.0
scikit-learn==1.4.0

# ML/최적화
lightgbm==4.3.0
xgboost==2.0.3
catboost==1.2.2
optuna==3.5.0
hmmlearn==0.3.0

# 트레이딩
binance-futures-connector==3.5.0
ccxt==4.2.0
websockets==12.0
aiohttp==3.9.0

# 성능
numba==0.59.0
joblib==1.3.2
tqdm==4.66.0

# 유틸리티
pyyaml==6.0.1
pydantic==2.5.0
rich==13.7.0

# 테스팅
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0